requirements.txt
discord.py
pyinstaller
mysql-connector-python
python-dotenv
sqlalchemy